+++ AndroidWrapperProject/app/src/main/java/io/branch/unity/BranchUnityApp.java
@@ -1,0 +1,17 @@
+package io.branch.unity;
+
+import android.app.Application;
+import io.branch.referral.Branch;
+
+/**
+ * Created by antonarhunou on 1/9/18.
+ */
+
+public class BranchUnityApp extends Application {
+    public void onCreate() {
+        super.onCreate();
+
+        Branch.getAutoInstance(this.getApplicationContext());
+        Branch.disableInstantDeepLinking(true);
+    }
+}