+++ BranchUnityTestBed/Assets/Branch/Demo/Scripts/BranchDemo.cs
@@ -57,6 +57,22 @@ public class BranchDemo : MonoBehaviour {
 
 			Debug.Log("Universal Object: " + universalObject.ToJsonString());
 			Debug.Log("Link Properties: " + linkProperties.ToJsonString());
+
+			BranchEvent e = new BranchEvent ("MY_CUSTOM_EVENT");
+//			BranchEvent e = new BranchEvent (BranchEventType.COMPLETE_REGISTRATION);
+
+			e.SetAffiliation("my_affilation");
+			e.SetCoupon("my_coupon");
+			e.SetCurrency(BranchCurrencyType.USD);
+			e.SetTax(10.0f);
+			e.SetRevenue(100.0f);
+			e.SetShipping(1000.0f);
+			e.SetDescription("my_description");
+			e.SetSearchQuery("my_search_query");
+			e.AddCustomData("custom_data_key01", "custom_data_value01");
+			e.AddContentItem(universalObject);
+
+			Branch.sendEvent (e);
 		}
 	}
 
@@ -82,7 +98,30 @@ public class BranchDemo : MonoBehaviour {
 			universalObject.title = "id12345 title";
 			universalObject.contentDescription = "My awesome piece of content!";
 			universalObject.imageUrl = "https://s3-us-west-1.amazonaws.com/branchhost/mosaic_og.png";
-			universalObject.metadata.Add("foo", "bar");
+			universalObject.contentIndexMode = 0;
+			universalObject.localIndexMode = 0;
+			universalObject.expirationDate = new DateTime(2020, 12, 30);
+			universalObject.keywords.Add("keyword01");
+			universalObject.keywords.Add("keyword02");
+
+			universalObject.metadata.contentSchema = BranchContentSchema.COMMERCE_BUSINESS;
+			universalObject.metadata.quantity = 100f;
+			universalObject.metadata.price = 1000f;
+			universalObject.metadata.currencyType = BranchCurrencyType.USD;
+			universalObject.metadata.sku = "my_sku";
+			universalObject.metadata.productName = "my_product_name";
+			universalObject.metadata.productBrand = "my_product_brand";
+			universalObject.metadata.productCategory = BranchProductCategory.BUSINESS_AND_INDUSTRIAL;
+			universalObject.metadata.condition = BranchCondition.EXCELLENT;
+			universalObject.metadata.productVariant = "my_product_variant";
+
+			universalObject.metadata.setAddress("my_street", "my_city", "my_region", "my_comuntry", "my_postal_code");
+			universalObject.metadata.setLocation(40.0f, 40.0f);
+			universalObject.metadata.setRating(4.0f, 5.0f, 10);
+
+			universalObject.metadata.AddImageCaption("https://s3-us-west-1.amazonaws.com/branchhost/mosaic_og.png");
+			universalObject.metadata.AddCustomMetadata("foo", "bar");
+
 
 			// register a view to add to the index
 			Branch.registerView(universalObject);
@@ -93,7 +132,7 @@ public class BranchDemo : MonoBehaviour {
 			linkProperties.feature = "sharing";
 			linkProperties.channel = "facebook";
 			linkProperties.controlParams.Add("$desktop_url", "http://example.com");
-			
+
 			Branch.getShortURL(universalObject, linkProperties, (url, error) => {
 				if (error != null) {
 					Debug.LogError("Branch.getShortURL failed: " + error);
@@ -175,7 +214,29 @@ public class BranchDemo : MonoBehaviour {
 				universalObject.title = "id12345 title";
 				universalObject.contentDescription = "My awesome piece of content!";
 				universalObject.imageUrl = "https://s3-us-west-1.amazonaws.com/branchhost/mosaic_og.png";
-				universalObject.metadata.Add("foo", "bar");
+				universalObject.contentIndexMode = 0;
+				universalObject.localIndexMode = 0;
+				universalObject.expirationDate = new DateTime(2020, 12, 30);
+				universalObject.keywords.Add("keyword01");
+				universalObject.keywords.Add("keyword02");
+
+				universalObject.metadata.contentSchema = BranchContentSchema.COMMERCE_BUSINESS;
+				universalObject.metadata.quantity = 100f;
+				universalObject.metadata.price = 1000f;
+				universalObject.metadata.currencyType = BranchCurrencyType.USD;
+				universalObject.metadata.sku = "my_sku";
+				universalObject.metadata.productName = "my_product_name";
+				universalObject.metadata.productBrand = "my_product_brand";
+				universalObject.metadata.productCategory = BranchProductCategory.BUSINESS_AND_INDUSTRIAL;
+				universalObject.metadata.condition = BranchCondition.EXCELLENT;
+				universalObject.metadata.productVariant = "my_product_variant";
+
+				universalObject.metadata.setAddress("my_street", "my_city", "my_region", "my_comuntry", "my_postal_code");
+				universalObject.metadata.setLocation(40.0f, 40.0f);
+				universalObject.metadata.setRating(4.0f, 5.0f, 10);
+
+				universalObject.metadata.AddImageCaption("https://s3-us-west-1.amazonaws.com/branchhost/mosaic_og.png");
+				universalObject.metadata.AddCustomMetadata("foo", "bar");
 
 				// register a view to add to the index
 				Branch.registerView(universalObject);
@@ -228,7 +289,7 @@ public class BranchDemo : MonoBehaviour {
 
 	public void OnBtn_RegisterForSpotlight() {
 		if (universalObject != null) {
-			universalObject.metadata.Add("deeplink_text", "This link was generated for Spotlight registration");
+			universalObject.metadata.AddCustomMetadata("deeplink_text", "This link was generated for Spotlight registration");
 			Branch.listOnSpotlight(universalObject);
 		}
 	}